Added a check when identifying vertices ensuring vertex is far enough away
Cases exist where the first vertex (i.e. on the top row) is not really at
the extremes of the marker but further towards the middle of an edge.  This
is because the top edge looks a little like a staircase when thresholded.
This commit goes some way towards combatting this issue -- it's not perfect
yet.