Implemented grid code recovery from the image using Hamming marker design
This code ties in with tools/markergen.py in that it decodes the markers it
produces.  It checks all 4 rotations of a koki_grid_t to find the right one.
This commit adds functionality for decoding a koki_grid_t only.

Currently, CRCs are *NOT* checked, any code is accepted.  CRC functionality
required for that to work.

Ditz-issue: 7aec2fd317bbb2e3394e4401425746092d29cdb8