Implemented local adaptive thresholding to handle varying brighness in an image
No matter what the lighting is like* across the image, this thresholding method
should be able to find all of the marker borders without trouble.  Note that this
method is not suitable for all thresholding requirements: because of its adaptive
nature, large black regions can have their centre thresholded white due to the lack
of variation in colour in that region.  Increasing the window size could resolve
this, but with the penalty of _a lot_ more computation.

*to a reasonable extent.

Ditz-issue: 4e4389f5e968f009ebcdc1e0849b89056d06db60