Scan left, right, left... towards centre of clip region instead of left to right
The dodgy situation where the first vertex was on the RHS of the clip region is
now resolved.  Because of the aliasing nature of thresholding, there could be a
'step' at the right hand side of the marker giving a false vertex.  Scanning
towards the centre appears to have significantly improved this situation.

Ditz-issue: a3ce92c9c4d3f0fc472107f4aa9af7bbb7ab5538