Fix: ensure warped clip region has a width and height > 0
This was one of those bugs that was incredibly difficult to reproduce.
I believe the known bug where OpenCV perspective unwarping threw an
error message and died was caused by this.  Such a situation can
arise when certain shapes are misidentified as quads, PCA is applied,
and the resultant vertices have no X and/or Y relative spacing, i.e.
all X values are n, or all Y values are m, for example.