Rejuggled markergen.py script to be more generic, allow 4-up output, and short desc. opt.
The flags --4up and --desc can be used to output 4 (in a square) on the page and set a short
description for the marker, respectively.  The short description will be added to the end of
the text at the bottom of the marker and can be used to prompt the user.