Added a higher level function to allow different marker widths depending on code
A function:  'f(code) --> marker width'  can be specified which returns the width
of the marker when given the code.  Having this functionality makes it easy to
support multiple marker sizes without going deeper into libkoki.

Doing so makes it easy for markers to have different sizes depending on their code