Fix: ensure quad finding recursion stops if num. found > 4
I believe this commit fixes a once-in-a-blue-moon segfault after
observing one of the points arrays being over filled and messing
up the stack.  When a contour is of some rather contorted shape,
the recursion can go on for a while, finding more than 10
vertices on the recursive calls for the first contour segment.
Not good.