Fix: slight trig. issue with vertical bearing calculation
bearing.y is calculated as the angle in the x-z plane, as expected.
bearing.x was being calculated in the y-z plane, which is incorrect.
When viewed as 2 sequential rotations, the resultant vector would be
incorrect.  The y angle that we actually want to know is the angle
between the x-z plane and r, the vector.  This small tweak corrects
this.

Note that, typically, this angle is presented as the angle between the
y-axis and r.  For most robotics applications, however, the value used
above is more useful.