Use an integral image in calculating local averages for thresholding
The area summation is currently quite a hot piece of code.  This
patch, which introduces the use of an integral image, cools it down a
little.

It takes koki_threshold_adaptive from taking 59% of CPU time down to
50%.  Not a greatly significant reduction, but at least a step in the
right direction.  Presumably this reduction is not as much as hoped
because the integral image is a 32-bit image, so accesses to it take
longer.  However, there are fewer memory accesses with it, hence the
speedup.