Switch to using scons for most of the build
In order to move libkoki towards a sitation in which we can properly
install it, and perhaps do some prerequisite testing, we could move to
using autotools.  Autotools has quite a reputation, and I think
rightly so.  Looking around at the available options, scons seems like
it might be the best around.

This patch switches much of the libkoki build to use scons.  There are
some things that remain unconverted, including:

 - AUTHORS file generation: This has no prerequisites except for the
   git history, so will need a scons builder that interacts with git.
 - ditz stuff: Haven't found a scons builder for ditz, so this will
   also need a builder writing.