Add and use a label_t type for numbering labels
Currently, labels are represented as 16-bit numbers.  With large
images, more than 0xffff regions may need to be labelled.  For
example, by feeding a 3264 x 2448 pixel image from my compact camera
into marker_info, bad things happen -- like an infinite loop waiting
for a 16-bit integer to reach 80000...

This patch replaces the type used for label numbers with a label_t
typedef.  This patch keeps it as a uint16_t for the moment, as it is
likely to negatively affect performance for 'smaller' images.