Adaptively threshold unwarped markers
This patch significantly improves the reliability of the extraction of
codes from unwarped markers.  Unwarped markers were being thresholded
'globally' (i.e. one threshold was applied against all the pixels of
the image) with a threshold that had been calculated using a heuristic
that attempted to balance the number of white and black pixels.

Now, as with the thresholding of the whole input image, the threshold
applied to each pixel within the unwarped marker image is based upon
the pixels around it.

This was the result of the work initiated by this thread on
srobo-devel:
https://groups.google.com/d/topic/srobo-devel/wa4ERK_9no0/discussion

We didn't originally adaptively threshold unwarped markers because it
appeared to reduce performance.  This turned out to be due to a bug
that was fixed in 7cd671161b94.

On Chris's set of test images, marker identification is improved from
this:

    Results
    ========
    True positives  728
    True negatives  995
    False positives 0
    False negatives 267

    Precision 100.0%
    Recall    73.1658291457%
    Accuracy  86.5829145729%

to this:

    Results
    ========
    True positives  826
    True negatives  995
    False positives 0
    False negatives 169

    Precision 100.0%
    Recall    83.0150753769%
    Accuracy  91.5075376884%