Add the beginnings of a libkoki context type
There are certain things that we want to do that would significantly
benefit from a common context instance being passed through the image
processing chain.  The thing that's mainly motivating me to do this is
the implementation of some logging stuff that'll be able to log both
text and images.  This will hopefully make it easier to debug things
when libkoki doesn't find a marker that one wants etc.

This patch adds a couple of functions that create a (currently
completely empty) instance of a "koki_t".  The user will call this,
optionally configure it, then pass it through to other functions that
need it -- e.g. koki_find_markers.