Remove a division through rearrangement
koki_threshold_adaptive_pixel was performing some division.  On ARM
platforms this was responsible for about 10% of the processing time.
This patch rearranges the calculations in threshold_adaptive_pixel so
that instead of dividing then comparing, the other side is multiplied
and then compared.

This results in speedup on my BeagleBone and x86 PC.